Improved evaluate_performance tool.
- Added support for variaous plot setups (e.g. Run 2, Phase 2 etc.)
- Store ROC curves in JSON format
- Updated documentation including section on how to use the code
  within LCG or conda environments.
